plat: marvell: armada: a3k: add Turris Mox platform to solve reset issue
The Turris Mox board has a HW bug which causes reset to hang
unpredictably. This issue can be solved by putting the board in a
specific state before reset. We solve this from WTMI firmware running on
the Cortex-M3 secure coprocessor and this function is exposed via the
mailbox interface.

Add platform for Turris Mox and add a call to this reset handler from
a3700_system_reset.

Signed-off-by: Marek Behún <marek.behun@nic.cz>